[Hooks] Create branch protection system
This commit introduces the `protected_refs.txt` file. This file is
copied in all challenge `hooks/` folder and is queried by the
`update` hook. If the `ref` to be updated (pushed by the player) is
found in this file, the push is rejected, and no challenge tests
are run.

The purpose of this is to protect branch apart from `master` in a
per-challenge way. For instance, a challenge that is about merging
2 (or more) upstream branches is now possible (`master` and
`hotfix`).

By default, the file exists in `generic` (copied to all challenges)
and contains only `master` (previous behaviour). Creating such a file
in a challenge directory (e.g `challenge4/`), does overwrite the
file copied from `generic`.