[Hooks] Asciinema solution support
This commit checks in the challenges `hooks/` directory for
`asciinema.txt` file and presents it to the player, after they
complete the challenge.

This feature was requested by solvers, so they can get an idea
of "intended ways" and "best approaches" to tackle the challenges.

The feature does support multiple URLs as there are many git ways
to achieve the same goal.