[Generic] Add Code Checking in gitcourselib
This commit adds several stuff in the library. Specifically, the
`check_code` is implemented that accepts a commit and several
keyphrases and ensures the contents of the commit patch contain the
given keyphrases, in a per-file basis.

This is a functionality already implemented in the previous
challenges in a manual way, and now it is generalized and coded in
the `gitcourselib`.

Additioanlly, the `get_title_tag` extracts a commit tag (what is
inside the `[]` and returns it.

These values are used to identify commits that are supposed to
contain specific code.