[Automation] Heroku Release Github Workflow
This commit creates a Github Workflow that leverages the Github
Secrets (where a Heroku API key is stored) to automatically create
a new release on the Heroku App when new commits hit the `master`
branch.

The Make file is modified to be able to create `hooks/` directory
in all challenge repositories.